+++ interface/src/routes/statusbar.svelte
@@ -1,10 +1,39 @@
 <script lang="ts">
 	import { page } from '$app/stores';
 	import { telemetry } from '$lib/stores/telemetry';
+	import { openModal, closeModal } from 'svelte-modals';
+	import { user } from '$lib/stores/user';
+	import ConfirmDialog from '$lib/components/ConfirmDialog.svelte';
 	import CPU from '~icons/tabler/cpu';
 	import WiFiOff from '~icons/tabler/wifi-off';
 	import Hamburger from '~icons/tabler/menu-2';
+	import Power from '~icons/tabler/power';
+	import Cancel from '~icons/tabler/x';
 	import RssiIndicator from '$lib/components/RSSIIndicator.svelte';
+
+	async function postSleep() {
+		const response = await fetch('/rest/sleep', {
+			method: 'POST',
+			headers: {
+				Authorization: $page.data.features.security ? 'Bearer ' + $user.bearer_token : 'Basic'
+			}
+		});
+	}
+
+	function confirmSleep() {
+		openModal(ConfirmDialog, {
+			title: 'Confirm Power Down',
+			message: 'Are you sure you want to switch off the device?',
+			labels: {
+				cancel: { label: 'Abort', icon: Cancel },
+				confirm: { label: 'Switch Off', icon: Power }
+			},
+			onConfirm: () => {
+				closeModal();
+				postSleep();
+			}
+		});
+	}
 </script>
 
 <div class="navbar bg-base-300 sticky top-0 z-10 h-12 min-h-fit drop-shadow-lg lg:h-16">
@@ -28,7 +57,14 @@
 		{#if $telemetry.rssi.disconnected}
 			<WiFiOff class="h-9 w-9" />
 		{:else}
-			<RssiIndicator showDBm={true} rssi_dbm={$telemetry.rssi.rssi} class="h-9 w-9" />
+			<RssiIndicator showDBm={false} rssi_dbm={$telemetry.rssi.rssi} class="h-9 w-9" />
+		{/if}
+	</div>
+	<div class="flex-none">
+		{#if $page.data.features.sleep}
+			<button class="btn btn-square btn-ghost" on:click={confirmSleep}>
+				<Power class="text-error h-7 w-7" />
+			</button>
 		{/if}
 	</div>
 </div>