+++ lib/framework/Features.h
@@ -42,4 +42,9 @@
 #define FT_UPLOAD_FIRMWARE 0
 #endif
 
+// ESP32 sleep states off by default
+#ifndef FT_SLEEP
+#define FT_SLEEP 0
+#endif
+
 #endif