+++ lib/framework/FeaturesService.cpp
@@ -48,6 +48,11 @@ void FeaturesService::features(AsyncWebServerRequest *request)
 #else
     root["upload_firmware"] = false;
 #endif
+#if FT_ENABLED(FT_SLEEP)
+    root["sleep"] = true;
+#else
+    root["sleep"] = false;
+#endif
     response->setLength();
     request->send(response);
 }