+++ lib/framework/FeaturesService.h
@@ -19,7 +19,6 @@
 
 #include <WiFi.h>
 #include <AsyncTCP.h>
-
 #include <ArduinoJson.h>
 #include <AsyncJson.h>
 #include <ESPAsyncWebServer.h>