+++ lib/framework/OTASettingsService.cpp
@@ -56,7 +56,6 @@ void OTASettingsService::configureArduinoOTA()
         _arduinoOTA->setPort(_state.port);
         _arduinoOTA->setPassword(_state.password.c_str());
         _arduinoOTA->setMdnsEnabled(false);
-        MDNS.enableArduino(_state.port, (_state.password.length() > 0));
         _arduinoOTA->onStart([]()
                              { Serial.println(F("Starting")); });
         _arduinoOTA->onEnd([]()
@@ -76,7 +75,9 @@ void OTASettingsService::configureArduinoOTA()
         Serial.println(F("Receive Failed"));
       else if (error == OTA_END_ERROR)
         Serial.println(F("End Failed")); });
+
         _arduinoOTA->begin();
+        MDNS.enableArduino(_state.port, (_state.password.length() > 0));
     }
 }